Create close popup module
