+++ src/modules/closePopup.js
@@ -1,0 +1,9 @@
+const closePopup = () => {
+    document.querySelector('.close').addEventListener('click', () => {
+      popup.style.display = 'none';
+      popup.innerHTML = '';
+      document.body.style.overflow = 'visible';
+    });
+  };
+
+  export default closePopup;
